[CRT] - Implement MSVC versions of call_copy_ctor, call_dtor, continue_after_catch, __CxxFrameHandler - Fix amd64 versions of _control87 and _statusfp - Fix MSVC and amd64 warnings
svn path=/trunk/; revision=53713

+++ reactos/lib/sdk/crt/conio/cputs.c
@@ -26,7 +26,7 @@ int _cputs(const char *_str)
   UNLOCK_CONSOLE;
   return retval;
 #else
-  int len = strlen(_str);
+  int len = (int)strlen(_str);
   DWORD written = 0;
   if (!WriteFile( fdesc[stdout->_file].hFile ,_str,len,&written,NULL))
     return -1;
